Move config utils into appropriate java module
