+++ libraries/config/src/test/java/ConfigUtilsTest.java
@@ -22,7 +22,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import org.dataportabilityproject.security.ConfigUtils;
+import org.dataportabilityproject.config.ConfigUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.runners.MockitoJUnitRunner;
