+++ portability-gateway/build.gradle
@@ -36,6 +36,7 @@ jar {
 }
 
 dependencies {
+    compile project(':libraries:config')
     compile project(':libraries:security')
     compile project(':portability-spi-service')
     compile project(':portability-spi-cloud')
