+++ portability-gateway/src/main/java/org/dataportabilityproject/gateway/reference/ReferenceApiModule.java
@@ -31,9 +31,9 @@ import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
 import javax.inject.Named;
+import org.dataportabilityproject.config.ConfigUtils;
 import org.dataportabilityproject.gateway.ApiSettings;
 import org.dataportabilityproject.security.AsymmetricKeyGenerator;
-import org.dataportabilityproject.security.ConfigUtils;
 import org.dataportabilityproject.security.RsaSymmetricKeyGenerator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
