+++ portability-worker/build.gradle
@@ -50,6 +50,7 @@ dependencies {
     compile project(':portability-spi-service')
     compile project(':portability-spi-cloud')
     compile project(':portability-spi-transfer')
+    compile project(':libraries:config')
     compile project(':libraries:security')
 
     compile("com.flickr4java:flickr4java:${flickrVersion}")
